Fix minor compiler warnings

+++ utils/usdData.h
@@ -364,7 +364,7 @@ struct USDFFUTILS_API UsdData
     std::string doc;
     PXR_NS::VtDictionary metadata;
     bool hasAnimations = false;
-    float minTime = std::numeric_limits<int>::max();
+    float minTime = std::numeric_limits<float>::max();
     float maxTime = 0;
     std::string animationName;
     double timeCodesPerSecond = 24;
